pyproject.toml
[tool.poetry]
name = "refactor"
version = "0.1.0"
description = ""
authors = ["Roland Parnaso <roland.parnaso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
aws-cdk-lib = "^2.102.0"
constructs = "^10.3.0"
python-dotenv = "^1.0.0"
boto3 = "^1.28.70"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
